添加seek 代码代码健壮性处理

+++ qplayer/src/main/cpp/q_player.h
@@ -64,7 +64,7 @@ public:
 
     void setRenderCallback(RenderingCallBack renderingCallBack);
 
-    int getDuration();
+    int64_t getDuration();
 
     void seek(int play_value);
